refactor(dialog-color): simplified dialog color
