+++ src/ui/units/wizard/components/Dialogs/DialogField/components/ButtonColorDialog/ButtonColorDialog.tsx
@@ -78,7 +78,6 @@ export const ButtonColorDialog: React.FC<ButtonColorDialogProps> = (
                         extra: extraSettings,
                         onApply: onApplyDialogColor,
                         isColorModeChangeAvailable: false,
-                        isMultipleColorsSupported: false,
                     }),
                 )
             }
